+++ src/esss_fix_format/cli.py
@@ -4,9 +4,11 @@ import os
 import re
 import subprocess
 import sys
+from typing import Optional, Tuple
 
 import click
 import pydevf
+from pathlib import Path
 
 CPP_PATTERNS = {
     '*.cpp',
@@ -54,6 +56,23 @@ def should_format(filename):
     if any(fnmatch(os.path.basename(filename), p) for p in PATTERNS):
         return True, ''
     return False, 'Unknown file type'
+
+
+def find_black_config(files_or_directories) -> Optional[Path]:
+    """
+    Searches for a valid pyproject.toml file based on the list of files/directories given.
+
+    We find the common ancestor of all files/directories given, then search from there
+    upwards for a "pyproject.toml" file with a "[tool.black]" section.
+    """
+    if not files_or_directories:
+        return None
+    common = Path(os.path.commonpath(files_or_directories))
+    for p in ([common] + list(common.parents)):
+        fn = p / 'pyproject.toml'
+        if fn.is_file() and '[tool.black]' in fn.read_text(encoding='UTF-8'):
+            return fn
+    return None
 
 
 # caches which directories have the `.clang-format` file, *in or above it*, to avoid hitting the
@@ -263,13 +282,14 @@ def _process_file(filename, check, format_code):
         if new_contents is None:
             new_contents = original_contents
 
-        try:
-            # Pass code formatter.
-            new_contents = format_code(new_contents)
-        except Exception as e:
-            error_msg = f'Error formatting code: {e}'
-            click.secho(error_msg, fg='red')
-            errors.append(error_msg)
+        if format_code is not None:
+            try:
+                # Pass code formatter.
+                new_contents = format_code(new_contents)
+            except Exception as e:
+                error_msg = f'Error formatting code: {e}'
+                click.secho(error_msg, fg='red')
+                errors.append(error_msg)
 
         if new_contents and (new_contents[0] == codecs.BOM_UTF8.decode('UTF-8')):
             msg = ': ERROR python file should not have a BOM.'
@@ -291,7 +311,35 @@ def _process_file(filename, check, format_code):
     return changed, errors, formatter
 
 
-def _main(files_or_directories, check, stdin, commit, format_code):
+def run_black_on_python_files(files, check) -> Tuple[bool, bool]:
+    """
+    Runs black on the given files (checking or formatting).
+
+    Black's output is shown directly to the user, so even in events of errors it is
+    expected that the users sees the error directly.
+
+    :return: a pair (formatted, error)
+    """
+    py_files = [x for x in files if x.suffix == '.py']
+    error = False
+    formatted = False
+    if py_files:
+        if check:
+            click.secho('Checking black...', fg='cyan')
+        else:
+            click.secho('Running black...', fg='cyan')
+        args = ['black']
+        if check:
+            args.append('--check')
+        args.extend(str(x) for x in py_files)
+        status = subprocess.call(args)
+        error = not check and status != 0
+        formatted = check and status == 1
+
+    return formatted, error
+
+
+def _main(files_or_directories, check, stdin, commit, pydevf_format_func):
     if stdin:
         files = [x.strip() for x in click.get_text_stream('stdin').readlines()]
     elif commit:
@@ -304,10 +352,23 @@ def _main(files_or_directories, check, stdin, commit, format_code):
                     files.extend(os.path.join(root, n) for n in names if should_format(n))
             else:
                 files.append(file_or_dir)
-    changed_files = []
+
+    files = [Path(x) for x in files]
     errors = []
+
+    black_config = find_black_config(files)
+    formatted_by_black = False
+    if black_config:
+        # skip pydevf formatter
+        pydevf_format_func = None
+        formatted_by_black, error = run_black_on_python_files(files, check)
+        if error:
+            errors.append('Error formatting black (see console)')
+
+    changed_files = []
     for filename in files:
-        changed, new_errors, formatter = _process_file(filename, check, format_code)
+        filename = str(filename)
+        changed, new_errors, formatter = _process_file(filename, check, pydevf_format_func)
         errors.extend(new_errors)
         if changed:
             changed_files.append(filename)
@@ -329,7 +390,7 @@ def _main(files_or_directories, check, stdin, commit, format_code):
         for error_msg in errors:
             click.secho(error_msg, fg='red')
         sys.exit(1)
-    if check and changed_files:
+    if check and (changed_files or formatted_by_black):
         click.secho('')
         click.secho(banner('failed checks'), fg='yellow')
         for filename in changed_files: